pyproject.toml
[project]
name = "researcharena"
version = "0.1.0"
description = "Benchmark harness for CLI agent autonomous research"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
paperreview = ["playwright>=1.40"]
dev = ["pytest", "ruff"]

[project.scripts]
researcharena = "researcharena.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
